+++ src/parseIntBinary.ts
@@ -1,0 +1,12 @@
+import { parseInt } from "./parseInt";
+
+/**
+ * Parses the given string as a binary number.
+ *
+ * @category Number
+ * @example
+ * ```typescript
+ * parseIntBinary("100"); // 0b100
+ * ```
+ */
+export const parseIntBinary = parseInt(2);