+++ src/parseIntDecimal.ts
@@ -1,0 +1,12 @@
+import { parseInt } from "./parseInt";
+
+/**
+ * Parses the given string as a decimal number.
+ *
+ * @category Number
+ * @example
+ * ```typescript
+ * parseIntDecimal("100"); // 100
+ * ```
+ */
+export const parseIntDecimal = parseInt(10);