+++ src/parseIntHexadecimal.ts
@@ -1,0 +1,12 @@
+import { parseInt } from "./parseInt";
+
+/**
+ * Parses the given string as a hexadecimal number.
+ *
+ * @category Number
+ * @example
+ * ```typescript
+ * parseIntHexadecimal("100"); // 0x100
+ * ```
+ */
+export const parseIntHexadecimal = parseInt(16);