+++ src/parseIntOctal.ts
@@ -1,0 +1,12 @@
+import { parseInt } from "./parseInt";
+
+/**
+ * Parses the given string as a octal number.
+ *
+ * @category Number
+ * @example
+ * ```typescript
+ * parseIntOctal("100"); // 0o100
+ * ```
+ */
+export const parseIntOctal = parseInt(8);